swap class rename + repetitive code removal
