+++ apps/charge-smart-wallets-service/src/common/services/user-op-factory.service.ts
@@ -1,6 +1,6 @@
+import { UserOpParser } from '@app/common/services/user-op-parser.service'
 import { UserOp } from '@app/smart-wallets-service/data-layer/interfaces/user-op.interface'
 import { Injectable } from '@nestjs/common'
-import { UserOpParser } from '@app/smart-wallets-service/common/services/user-op-parser.service'
 
 @Injectable()
 export class UserOpFactory {
