+++ libs/common/src/services/user-op-parser.service.ts
@@ -1,11 +1,11 @@
-import { decodeWithCalldata, sigHashFromCalldata } from '@app/common/utils/dtools/decodeBySigHash'
+import { decodeWithSelector } from '@app/common/utils/calldata-decoder/decoder'
 import { Injectable } from '@nestjs/common'
 import { BigNumber } from 'nestjs-ethers'
 
 @Injectable()
 export class UserOpParser {
   async decodeCalldata (callData: string) {
-    return decodeWithCalldata(sigHashFromCalldata(callData), callData)
+    return decodeWithSelector({ calldata: callData })
   }
 
   private transformArray (input) {
@@ -32,33 +32,33 @@ export class UserOpParser {
         return {
           targetAddress: call.targetAddress,
           value: call.value.toString(),
-          callData: decodedCallData[0].decoded,
-          name: decodedCallData[0].fragment.name
+          callData: decodedCallData.args,
+          name: decodedCallData.name || decodedCallData.fragment?.name
         }
       }
     }))
   }
 
   async parseCallData (callData: string) {
-    const decodeResults = await this.decodeCalldata(
-      callData
-    )
+    const decodeResults = await this.decodeCalldata(callData)
     if (!decodeResults) {
       throw new Error('Signature is wrong or undefined')
     }
 
-    const { decoded, fragment } = decodeResults[0]
+    const { args, name, fragment } = decodeResults
+    const fragmentName = name || fragment?.name
+
     const calls = []
-    if (fragment.name === 'executeBatch') {
-      const items = this.transformArray(decoded)
+    if (fragmentName === 'executeBatch') {
+      const items = this.transformArray(args)
       calls.push(...items)
     } else {
-      const items = this.transformArray([[decoded[0]], [decoded[1]], [decoded[2]]])
+      const items = this.transformArray([[args[0]], [args[1]], [args[2]]])
       calls.push(...items)
     }
 
     const targetFunctions = await this.getTargetFunction(calls)
 
-    return { name: fragment.name, targetFunctions, calls }
+    return { name: fragmentName, targetFunctions }
   }
 }
